+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 * Update to orient 3.0.25
 * Disable hibernate validator JPA support only if orient object support used
     (only in this case misactivation happen).
+* (breaking) Removed `HasOrientServerConfiguration` interface. Instead, 
+    configuration provider must be registered: `new OrientServerBundle<MyConfig>(MyConfig::getOrientConfig)`    
 
 ### 2.2.0 (2018-10-20)
 * Orient 3 (3.0.9) compatibility